iio: adrv9002: add multi chip sync support
The MCS support is given via a device attribute. The interface is very
simple:

Userspace needs to write to the attr ('multi_chip_sync') to set all
channels in MCS_READY state and start MCS. Then, the driver
automatically polls on the device radio state to know when MCS done. The
drivers polls for 10s before returning ETIMEDOUT (we can also break
early in case of an API error).

Given the above, userspace needs to make sure the pulses are given while
the driver is polling. We might need to wait a bit before starting
sending pulses (so the device is in the correct state).

Note that there are still two major failuree with MCS:

 1. We can only run MCS once. That's why the driver is not allowing
multiple runs;
 2. We need to force the carrier loGenOptimization value since the device
returns an invalid value when calling adi_adrv9001_Radio_Carrier_Inspect().

Signed-off-by: Nuno Sá <nuno.sa@analog.com>